requirements.txt
# Archivo: requirements.txt

# Biblioteca para la integración con Telegram
python-telegram-bot==13.15

# ORM para la gestión de la base de datos (en este ejemplo, SQLAlchemy)
SQLAlchemy==1.4.46

# Cliente de Mercado Pago para la integración del Checkout Pro
mercadopago==2.2.0

# Biblioteca para realizar peticiones HTTP (útil en algunas integraciones)
requests==2.28.1
